backend/requirements-api.txt
# ───── API Dependencies ─────
fastapi>=0.108.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1
gunicorn>=21.2.0
requests>=2.31.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ───── JWT & Security ─────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0

# ───── CORS y Middleware ─────
starlette>=0.36.0
python-dotenv>=1.0.0
structlog>=24.1.0
